Should give the square, cube and factorial of number user inputs.
It does not, however do that at the moment, but does
exactly that when the js and css are added in the html
file, while the links to the js and css files are
commented out.

I still do not understand why it does that, but I
will be back.